refactor: useCreateArtworkAlert hook to reuse functionality
