+++ src/app/Scenes/Artwork/Components/ArtworkScreenHeaderCreateAlert.tsx
@@ -1,21 +1,9 @@
-import {
-  ActionType,
-  ContextModule,
-  OwnerType,
-  ScreenOwnerType,
-  TappedCreateAlert,
-} from "@artsy/cohesion"
+import { ActionType, ContextModule, ScreenOwnerType, TappedCreateAlert } from "@artsy/cohesion"
 import { BellIcon, Button } from "@artsy/palette-mobile"
 import { ArtworkScreenHeaderCreateAlert_artwork$data } from "__generated__/ArtworkScreenHeaderCreateAlert_artwork.graphql"
 import { CreateSavedSearchModal } from "app/Components/Artist/ArtistArtworks/CreateSavedSearchModal"
-import { Aggregations } from "app/Components/ArtworkFilter/ArtworkFilterHelpers"
-import {
-  SavedSearchEntity,
-  SavedSearchEntityArtist,
-  SearchCriteriaAttributes,
-} from "app/Components/ArtworkFilter/SavedSearch/types"
-import { compact, isEmpty } from "lodash"
-import { useState } from "react"
+import { useCreateArtworkAlert } from "app/utils/hooks/useCreateArtworkAlert"
+import { isEmpty } from "lodash"
 import { createFragmentContainer, graphql } from "react-relay"
 import { useTracking } from "react-tracking"
 
@@ -28,63 +16,21 @@ const ArtworkScreenHeaderCreateAlert: React.FC<ArtworkScreenHeaderCreateAlertPro
 }) => {
   const { isForSale } = artwork
   const { trackEvent } = useTracking()
-  const [isCreateAlertModalVisible, setIsCreateAlertModalVisible] = useState(false)
-
   const hasArtists = !isEmpty(artwork.artists) && artwork.artists!.length > 0
+
+  const { entity, attributes, aggregations, showModal, closeModal, isCreateAlertModalVisible } =
+    useCreateArtworkAlert(artwork)
 
   if (!hasArtists) {
     return null
   }
 
-  let aggregations: Aggregations = []
-  let additionalGeneIDs: string[] = []
-
-  const artists = compact(artwork?.artists)
-  const attributionClass = compact([artwork?.attributionClass?.internalID])
-  const formattedArtists: SavedSearchEntityArtist[] = artists.map((artist) => ({
-    id: artist.internalID,
-    name: artist.name!,
-  }))
-
-  const entity: SavedSearchEntity = {
-    placeholder: `Artworks like: ${artwork?.title!}`,
-    artists: formattedArtists,
-    owner: {
-      type: OwnerType.artwork,
-      id: artwork?.internalID ?? "",
-      slug: artwork?.slug ?? "",
-    },
-  }
-
-  if (artwork?.mediumType?.filterGene?.name && artwork?.mediumType?.filterGene.slug) {
-    additionalGeneIDs = [artwork.mediumType.filterGene.slug]
-    aggregations = [
-      {
-        slice: "MEDIUM",
-        counts: [
-          {
-            name: artwork.mediumType.filterGene.name,
-            value: artwork.mediumType.filterGene.slug,
-            count: 0,
-          },
-        ],
-      },
-    ]
-  }
-
-  const attributes: SearchCriteriaAttributes = {
-    artistIDs: formattedArtists.map((artist) => artist.id),
-    attributionClass,
-    additionalGeneIDs,
-  }
-
   const handleCreateAlertPress = () => {
-    setIsCreateAlertModalVisible(true)
-
+    showModal()
     const event = tracks.tappedCreateAlert({
-      ownerId: entity.owner.id,
-      ownerType: entity.owner.type,
-      ownerSlug: entity.owner.slug,
+      ownerId: entity?.owner.id!,
+      ownerType: entity?.owner.type!,
+      ownerSlug: entity?.owner.slug!,
     })
 
     trackEvent(event)
@@ -103,10 +49,10 @@ const ArtworkScreenHeaderCreateAlert: React.FC<ArtworkScreenHeaderCreateAlertPro
       </Button>
       <CreateSavedSearchModal
         visible={isCreateAlertModalVisible}
-        entity={entity}
-        attributes={attributes}
-        aggregations={aggregations}
-        closeModal={() => setIsCreateAlertModalVisible(false)}
+        entity={entity!}
+        attributes={attributes!}
+        aggregations={aggregations!}
+        closeModal={() => closeModal()}
       />
     </>
   )
